Extract CompositeCommand to Commands::Composite
